excel_dashboard_proyectos/water-program-dashboard/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
